backend: fix composeUp for nerdctl
Unfortunately, we need to handle the paths differently on Windows (since we
don't need to copy files into the VM there).  This includes a refactoring
so that all the prep for compose is done in a helper method, so that we
don't need to repeat the platform-specific setup for each method.

Signed-off-by: Mark Yen <mark.yen@suse.com>