backend: WSL: Copy files in/out of the VM from the VM
The WSL 9p/SMB shares seem to be intermittently broken when we copy things
around (the copy call fails, but retries work).  Attempt to work around
the issue by doing the copy from within the Linux VM (so we use DrvFS
instead, which should also be 9p-based).  This seems to sometimes help.

Signed-off-by: Mark Yen <mark.yen@suse.com>